refactor: 자료형이 int인 bonusNumber를 포장해주어 BonusNumber 객체 생성
